4090 11 11

+++ mlh/defenses/membership_inference/DP.py
@@ -141,7 +141,7 @@ class TrainTargetDPSGD(TrainTargetNormalLoss):
             self.scheduler.step()
             if e == self.epochs:
                 log_dict = {'Loss Type' : self.args.loss_type,"Train Epoch" : e, "Total Sample": len(train_loader.dataset),
-                            "Train Acc": train_acc, "Test Acc": test_acc, "Loss": np.mean(losses), "Total Time" : time.time() - t_start}
+                            "Train Acc": train_acc, "Test Acc": test_acc, "Loss": float(np.mean(losses)), "Total Time" : time.time() - t_start}
                 save_dict_to_yaml(log_dict, f'{self.log_path}/train_log.yaml')
                 
         try:
